Separate scripting into its own package.
Bumps some dependencies.
