+++ OAT.Scripting/OAT.Scripting.csproj
@@ -1,0 +1,34 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFrameworks>netstandard2.0;netstandard2.1;net6.0</TargetFrameworks>
+		<LangVersion>10.0</LangVersion>
+		<Nullable>Enable</Nullable>
+		<RootNamespace>Microsoft.CST.OAT</RootNamespace>
+		<Version>0.0.0-placeholder</Version>
+		<Company>Microsoft</Company>
+		<Authors>Microsoft</Authors>
+		<copyright>© Microsoft Corporation. All rights reserved.</copyright>
+		<RepositoryType>GitHub</RepositoryType>
+		<RepositoryUrl>https://github.com/Microsoft/OAT</RepositoryUrl>
+		<GeneratePackageOnBuild>false</GeneratePackageOnBuild>
+		<Description>Adding this package enables the ScriptOperation to run scripts embedded in rules.</Description>
+		<PackageId>Microsoft.CST.OAT.Scripting</PackageId>
+		<PackageTags>Object Analysis Toolkit</PackageTags>
+		<PackageVersion>0.0.0-placeholder</PackageVersion>
+		<PackageProjectUrl>https://github.com/microsoft/OAT</PackageProjectUrl>
+		<PackageIcon>icon-128.png</PackageIcon>
+		<PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
+		<IncludeSymbols>true</IncludeSymbols>
+		<SymbolPackageFormat>snupkg</SymbolPackageFormat>
+	</PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="4.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\OAT\OAT.csproj" />
+  </ItemGroup>
+
+</Project>
