+++ OAT.sln
@@ -20,6 +20,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OAT.Blazor.Components", "OA
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OAT.Blazor", "OAT.Blazor\OAT.Blazor.csproj", "{A4DDA081-65F4-48E7-AE7C-0397A74A2607}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OAT.Scripting", "OAT.Scripting\OAT.Scripting.csproj", "{94343999-D64D-47CA-A0BC-AF4B19DABCCD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -50,6 +52,10 @@ Global
 		{A4DDA081-65F4-48E7-AE7C-0397A74A2607}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A4DDA081-65F4-48E7-AE7C-0397A74A2607}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A4DDA081-65F4-48E7-AE7C-0397A74A2607}.Release|Any CPU.Build.0 = Release|Any CPU
+		{94343999-D64D-47CA-A0BC-AF4B19DABCCD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{94343999-D64D-47CA-A0BC-AF4B19DABCCD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{94343999-D64D-47CA-A0BC-AF4B19DABCCD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{94343999-D64D-47CA-A0BC-AF4B19DABCCD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
