+++ OAT/Resources.resx
@@ -162,7 +162,6 @@
   <data name="Err_ClauseMissingScript_{0}{1}" xml:space="preserve">
     <value>Rule {0} Clause {1} is missing required Script Data for Script operation.</value>
   </data>
-  
   <data name="Err_ScriptingDisabled_{0}{1}" xml:space="preserve">
     <value>Rule {0} Clause {1} has Script but RunScripts is false.</value>
   </data>
@@ -226,5 +225,10 @@
   <data name="Err_NoOperation_{0}{1}" xml:space="preserve">
     <value>Rule {0}, Clause {1} is invalid as the Operator is set to NoOperation.</value>
   </data>
-    
+  <data name="Err_ErrorLoadingScriptingAssembly" xml:space="preserve">
+    <value>Failed to load the OAT.Scripting assembly to enable scripting.</value>
+  </data>
+  <data name="Err_ExceptionLoadingScriptingAssembly_{0}" xml:space="preserve">
+    <value>Exception hit while attempting to load OAT.Scripting assembly: {0}.</value>
+  </data>
 </root>
